Corre

+++ search_algorithms/0-linear.c
@@ -10,7 +10,7 @@
 */
 int linear_search(int *array, size_t size, int value)
 {
-	int i = 0;
+	size_t i = 0;
 
 	if (array == NULL)
 		return (-1);
@@ -19,7 +19,7 @@ int linear_search(int *array, size_t size, int value)
 	{
 		printf("Value checked array[%i] = [%i]", i, array[i]);
 		if (array[i] == value)
-			return (i);
+			return ((int)i);
 	}
 
 	return (-1);
